remove go1.6 support
