+++ README.md
@@ -120,7 +120,7 @@ $ go run main.go
 
 ## Prerequisite
 
-Now Gin requires Go 1.6 or later and Go 1.7 will be required soon.
+Now Gin requires Go 1.7 or later and Go 1.8 will be required soon.
 
 ## Quick start
